Bulk upload

+++ src/main/java/com/acft/acft/Controllers/AcftManagerServiceController.java
@@ -4,8 +4,10 @@ import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.URLConnection;
 import java.util.List;
 
@@ -82,7 +84,8 @@ public class AcftManagerServiceController {
 
     @GetMapping("/testGroup/getXlsxFile/{testGroupId}/{passcode}")
     public void exportXlsxFileForTestGroup(HttpServletRequest request, HttpServletResponse response, @PathVariable Long testGroupId, @PathVariable String passcode){
-        exportXlsxFileUtility(request, response, testGroupId, passcode);
+        File file = acftManagerService.getXlsxFileForTestGroupData(testGroupId, passcode);
+        exportXlsxFileUtility(request, response, file, true);
     }
 
     @GetMapping("/testGroup/getXlsxFile/{testGroupId}")
@@ -90,8 +93,41 @@ public class AcftManagerServiceController {
         exportXlsxFileForTestGroup(request, response, testGroupId, "");
     }
 
-    public void exportXlsxFileUtility(HttpServletRequest request, HttpServletResponse response, Long testGroupId, String passcode){
-        File file = acftManagerService.getXlsxFileForTestGroupData(testGroupId, passcode);
+    @GetMapping("/getBulkUploadTemplate")
+    public void getBulkUploadTemplate(HttpServletRequest request, HttpServletResponse response){
+        File file = acftManagerService.getBulkUploadTemplate();
+        exportXlsxFileUtility(request, response, file, false);
+    }
+
+    //Might not need the response parater
+    @PostMapping("/bulkUpload/{testGroupId}/{passcode}")
+    public boolean bulkUpload(HttpServletRequest request, HttpServletResponse response, @PathVariable Long testGroupId, @PathVariable String passcode){
+        String path = "src/main/resources/data/bulkUpload.xlsx";
+        File file = new File(path);
+        OutputStream outputStream;
+        try {
+            outputStream = new FileOutputStream(path);
+            FileCopyUtils.copy(request.getInputStream(), outputStream);
+            outputStream.close();
+        } catch (FileNotFoundException e) {
+            System.out.println(e.getMessage());
+            return false;
+        } catch (IOException e) {
+            System.out.println(e.getMessage());
+            return false;
+        }
+        file = new File(path);
+        return acftManagerService.instantiateBulkUploadData(file, testGroupId, passcode);
+        
+    }
+
+    @PostMapping("/bulkUpload/{testGroupId}")
+    public boolean bulkUpload(HttpServletRequest request, HttpServletResponse response, @PathVariable Long testGroupId){
+        return bulkUpload(request, response, testGroupId, "");
+    }
+
+    //Might not need the request parameter
+    public void exportXlsxFileUtility(HttpServletRequest request, HttpServletResponse response, File file, boolean delete){
         String mimeType = URLConnection.guessContentTypeFromName(file.getName());
         response.setContentType(mimeType);
         response.setHeader("Content-Disposition", String.format("inline; filename=\"" + file.getName() + "\""));
@@ -105,7 +141,7 @@ public class AcftManagerServiceController {
         } catch (IOException e){
             System.out.println(e.getMessage());
         }   
-        file.delete();
+        if (delete) file.delete();
     }
 
     @DeleteMapping("/deleteAll")
